+++ Cargo.toml
@@ -32,6 +32,7 @@ clap_complete = "3.1.1"
 clap_mangen = "0.1"
 config = "0.12.0"
 criterion = "0.3.5"
+dirs = "4.0.0"
 git2 = "0.14.2"
 hex = "0.4.3"
 indoc = "1.0.4"
