+++ lib/Cargo.toml
@@ -23,7 +23,6 @@ bytes = "1.1.0"
 byteorder = "1.4.3"
 chrono = "0.4.19"
 config = "0.12.0"
-dirs = "4.0.0"
 git2 = "0.14.2"
 hex = "0.4.3"
 itertools = "0.10.3"
